requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
py-cord==2.6.1
python-dotenv==1.2.1
setuptools==80.9.0
wheel==0.45.1
yarl==1.22.0
